Bugfix for Uninstall-Msi  on Windows PowerShell
This fixes two bugs:
- Uninstall-Msi not executing the uninstaller in Windows PowerShell 5.1
- WriteLog not being able to run because the non existant parameter 'Severity' is set instead of the parameter 'LogLevel'.

In Windows PowerShell the Start-Process Cmdlet doesn't support the -WhatIf parameter.
Start-Process is being passed a WhatIf parameter in the function Uninstall-Msi and couldn't be executed. To fix this I removed the -WhatIf parameter from the parameter list and wrapped the code that calls the
uninstaller in an if statement that calls '$PSCmdlet.ShouldProcess()'.